+++ packages/frontend/src/downstream/Search/Output/Error.tsx
@@ -1,5 +1,0 @@
-interface IError {
-  text: string;
-}
-
-export { IError };